Removed dependency on Context in History
